+++ src/pgrls/rules/__init__.py
@@ -111,6 +111,7 @@ def _build_default_registry() -> RuleRegistry:
     from pgrls.rules.sec024 import SEC024
     from pgrls.rules.sec025 import SEC025
     from pgrls.rules.sec026 import SEC026
+    from pgrls.rules.sec027 import SEC027
     from pgrls.rules.view001 import VIEW001
     from pgrls.rules.view002 import VIEW002
     from pgrls.rules.view003 import VIEW003
@@ -143,6 +144,7 @@ def _build_default_registry() -> RuleRegistry:
     registry.register(SEC024())
     registry.register(SEC025())
     registry.register(SEC026())
+    registry.register(SEC027())
     registry.register(PERF001())
     registry.register(PERF002())
     registry.register(PERF003())